pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Daniel <matsubus@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
"discord.py" = "^1.7.3"
black = "^22.3.0"
isort = "^5.10.1"
certifi = "^2021.10.8"
SQLAlchemy = "^1.4.35"
pydantic = "^1.9.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
db = "^0.1.1"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

